TG-48: Changed form to an especific form that calls shared form
