+++ src/Components/TimeSheets/List/index.module.css
@@ -1,0 +1,11 @@
+.container {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.LinkReset {
+  text-decoration: none;
+}
